www/requirements.txt
# requirements.txt

# Standard dependencies.
omegaconf
bson
pydantic
email-validator
stripe

# AWS dependencies.
aioboto3
boto3-stubs[cognito]

# Crypto dependencies
cryptography

# FastAPI dependencies.
fastapi[standard]
pillow
python-multipart
httpx
httpx-ws
python-jose[cryptography]
pytest-mock

# Deployment dependencies.
uvicorn[standard]

# Processing dependencies.
numpy<2
numpy-stl
trimesh

# Types
types-aioboto3[dynamodb, s3]

# Converting files
pybullet
mujoco
kscale-onshape-library[all]
